[checkpoint] Fix dynasmrt 4.0.1 compatibility
Change register_to_dynasm() return type from u32 to u8 to align with
dynasmrt 4.0.1's requirement for Into<u8> on dynamic register arguments.

This resolves compilation errors introduced by the breaking change in
dynasmrt 4.0.0 where dynamic registers now accept any value convertible
via Into<u8>. AArch64 register indices (0-31) fit safely within u8.

+++ src/assembler/mod.rs
@@ -262,9 +262,9 @@ impl Default for AArch64Assembler {
     }
 }
 
-fn register_to_dynasm(reg: Register) -> Result<u32, String> {
+fn register_to_dynasm(reg: Register) -> Result<u8, String> {
     reg.index()
-        .map(|idx| idx as u32)
+        .map(|idx| idx as u8)
         .ok_or_else(|| format!("Register {:?} not supported in dynasm encoding", reg))
 }
